+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Changelog
 * added axisSoftMin and axisSoftMax options for TimeSeries
 * Added support for Azure Data Explorer datasource plugin (https://github.com/grafana/azure-data-explorer-datasource)
 * Added ``sortBy`` parameter to Table panel
+* Added ``tooltipSort`` parameter to TimeSeries panel
 
 .. _`Bar_Chart`: https://grafana.com/docs/grafana/latest/panels-visualizations/visualizations/bar-chart/
 .. _`RateMetricAgg`: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-metrics-rate-aggregation.html